Skip rendering when dynamic range is zero
When range_min==range_max, the rendering equation divides by zero and the
resulting scaled channel image is all NaN. This ends up forcing the entire
image to black as the NaNs propagate through the compositing and get turned
into zeros when cast to uint8. This change just skips rendering and
compositing when the dynamic range is zero.