Fix threading issues in auto-threshold code
Previous implementation deadlocked on Windows. This change fixes that
and cleans up the code in general. Now uses ThreadPoolExecutor instead
of explicitly launching Threads. Even after the general cleanups I was
encountering deadlocks in np.dot on normal numpy arrays in a threaded
context -- switching to the "blis" BLAS implementation fixed that
(requirements.yml now contains an entry for this).